fix: include provided deps in module cache keys
resolveVersion now includes a sorted list of module names (from the
provided dependencies) in the module version cache key.

Fixes: https://github.com/garden-io/garden/issues/307
